Added screen to help user to determine the tempo of a piece.  The user taps on the down button and the metronome displays the tempo (rate) at which the user is tapping.
Because PebbleOS is *really* weak in the area of timing, I had to
break out the ARM manuals and configure a hardware timer (TIM5) to
perform the interval timing.  TIM5 is configured to count at 1kHz.  It
uses a prescaler of 32768, so I assume that the Pebble is running at
32MHz.

This code is seriously lame right now, but it works.